teams-lunch-bot/requirements.txt
# Bot Framework
botbuilder-core>=4.16.0
botbuilder-integration-aiohttp>=4.16.0

# Web server
aiohttp>=3.9.0

# Reuse existing scraper (parent project)
# These are also in ../requirements.txt
fastmcp>=2.0.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
openai>=1.61.0
google-genai>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
